refactor(vtree): overhauling vtree to match v2 structure (wip)
