+++ packages/vtree/core/modeling/operations/booleans/intersect.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const intersection = (...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'intersection', params: undefined }
+}
+
+module.exports = intersection
