+++ packages/vtree/core/modeling/operations/booleans/subract.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const difference = (...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'difference', params: undefined }
+}
+
+module.exports = difference
