+++ packages/vtree/core/modeling/operations/booleans/union.js
@@ -1,0 +1,8 @@
+const { flatten } = require('./arrays')
+
+const union = (...solids) => {
+  solids = flatten(toArray(solids))
+  return { children: solids, type: 'union', params: undefined }
+}
+
+module.exports = union
