+++ packages/vtree/core/modeling/operations/expansions/expand.js
@@ -1,0 +1,13 @@
+const { toArray } = require('./arrays')
+
+const contract = (radius, n, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'contract', radius, n }
+}
+
+const expand = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'expand', params }
+}
+
+module.exports = expand
