+++ packages/vtree/core/modeling/operations/extrusions/extrudeLinear.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const extrudeLinear = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'extrudeLinear', params }
+}
+
+module.exports = extrudeLinear
