+++ packages/vtree/core/modeling/operations/extrusions/extrudeRectangular.js
@@ -1,0 +1,9 @@
+const { toArray } = require('./arrays')
+
+const extrudeRectangular = (solids, params) => {
+  console.log('extrudeRectangular')
+  solids = toArray(solids)
+  return { children: solids, type: 'extrudeRectangular', params }
+}
+
+module.exports = extrudeRectangular
