+++ packages/vtree/core/modeling/operations/extrusions/extrudeRotate.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const extrudeRotate = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'extrudeRotate', params }
+}
+
+module.exports = extrudeRotate
