+++ packages/vtree/core/modeling/operations/hulls/hull.js
@@ -1,0 +1,7 @@
+const { toArray } = require('./arrays')
+
+const hull = (...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'hull', params: undefined }
+}
+module.exports = hull
