+++ packages/vtree/core/modeling/operations/hulls/hullChain.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const hullChain = (...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'hullChain', params: undefined }
+}
+
+module.exports = hullChain
