+++ packages/vtree/core/modeling/operations/transforms/mirror.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const mirror = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'mirror', params }
+}
+
+module.exports = mirror
