+++ packages/vtree/core/modeling/operations/transforms/rotate.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const rotate = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'rotate', params }
+}
+
+module.exports = rotate
