+++ packages/vtree/core/modeling/operations/transforms/scale.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const scale = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'scale', params }
+}
+
+module.exports = scale
