+++ packages/vtree/core/modeling/operations/transforms/translate.js
@@ -1,0 +1,8 @@
+const { toArray } = require('./arrays')
+
+const translate = (params, ...solids) => {
+  solids = toArray(solids)
+  return { children: solids, type: 'translate', params }
+}
+
+module.exports = translate
