+++ packages/vtree/core/modeling/primitives/cuboid.js
@@ -1,0 +1,13 @@
+const { toArray } = require('./arrays')
+
+const cuboid = params => {
+  /* const defaults = {
+      size: [1, 1, 1],
+      center: [true, true, true]
+    } */
+  const _params = Object.assign({}, params)
+
+  return Object.assign({}, _params, { type: 'cuboid' })
+}
+
+module.exports = cuboid
