+++ packages/vtree/core/modeling/primitives/cylinder.js
@@ -1,0 +1,7 @@
+const cylinder = params => {
+  const _params = Object.assign({}, params)
+
+  return Object.assign({}, _params, { type: 'cylinder' })
+}
+
+module.exports = cylinder
