+++ packages/vtree/core/modeling/primitives/ellipse.js
@@ -1,0 +1,6 @@
+const ellipse = (params) => {
+  const _params = params
+  return Object.assign({}, _params, { type: 'ellipse' })
+}
+
+module.exports = ellipse
